requirements.txt
moviepy==1.0.3
decorator==4.4.2
openai-whisper
opencv-python
requests
sentence-transformers
torch
transformers